Start simplifying remote implementation
