+++ packages/react-native/src/sync/stream/ReactNativeRemote.ts
@@ -1,12 +1,5 @@
 import { LogLevels, PowerSyncLogger } from '@powersync/common';
-import {
-  AbstractRemote,
-  AbstractRemoteOptions,
-  FetchImplementation,
-  FetchImplementationProvider,
-  RemoteConnector,
-  SyncStreamOptions
-} from '@powersync/shared-internals';
+import { AbstractRemote, FetchOptions, RemoteConnector } from '@powersync/shared-internals';
 import { Platform } from 'react-native';
 // @ts-expect-error
 import { TextDecoder } from 'text-encoding';
@@ -16,46 +9,47 @@ import { fetch } from 'react-native-fetch-api';
 
 export const STREAMING_POST_TIMEOUT_MS = 30_000;
 
-type ReactNativeFetchOptions = RequestInit & {
-  reactNative?: Record<string, unknown>;
-};
-
-/**
- * Directly imports fetch implementation from react-native-fetch-api.
- * This removes the requirement for the global `fetch` to be overridden by
- * a polyfill.
- */
-class ReactNativeFetchProvider extends FetchImplementationProvider {
-  getFetch(): FetchImplementation {
-    return fetch.bind(globalThis);
-  }
-}
-
 export class ReactNativeRemote extends AbstractRemote {
   constructor(
     protected connector: RemoteConnector,
-    protected logger: PowerSyncLogger,
-    options?: Partial<AbstractRemoteOptions>
+    logger: PowerSyncLogger
   ) {
-    super(connector, logger, {
-      ...(options ?? {}),
-      fetchImplementation: options?.fetchImplementation ?? new ReactNativeFetchProvider()
-    });
+    super(connector, logger);
   }
 
-  get fetch(): FetchImplementation {
-    const fetchImplementation = super.fetch;
-    return ((input, init) => {
-      const options = (init ?? {}) as ReactNativeFetchOptions;
+  protected async fetch(options: FetchOptions): Promise<Response> {
+    let timeout: unknown;
+    if (options.expectStreamingResponse) {
+      // @ts-expect-error https://github.com/react-native-community/fetch#enable-text-streaming
+      options.request.reactNative = {
+        /**
+         * The `react-native-fetch-api` polyfill provides streaming support via
+         * this non-standard flag
+         * https://github.com/react-native-community/fetch#enable-text-streaming
+         */
+        textStreaming: true
+      };
 
-      return fetchImplementation(input, {
-        ...options,
-        reactNative: {
-          ...(options.reactNative ?? {}),
-          textStreaming: true
-        }
-      } as RequestInit);
-    }) as FetchImplementation;
+      timeout =
+        Platform.OS == 'android'
+          ? setTimeout(() => {
+              this.logger.log({
+                level: LogLevels.warn,
+                message: `HTTP Streaming POST is taking longer than ${Math.ceil(
+                  STREAMING_POST_TIMEOUT_MS / 1000
+                )} seconds to resolve. If using a debug build, please ensure Flipper Network plugin is disabled.`
+              });
+            }, STREAMING_POST_TIMEOUT_MS)
+          : null;
+    }
+
+    try {
+      // Directly import the fetch implementation from react-native-fetch-api. This removes the requirement for the
+      // global `fetch` to be overridden by a polyfill.
+      return await fetch(options.resource, options.request);
+    } finally {
+      clearTimeout(timeout as any);
+    }
   }
 
   getUserAgent(): string {
@@ -74,39 +68,5 @@ export class ReactNativeRemote extends AbstractRemote {
   protected get supportsStreamingBinaryResponses(): boolean {
     // We have to pass textStreaming: true to get streamed responses at all, and those don't support binary data.
     return false;
-  }
-
-  protected async fetchStreamRaw(options: SyncStreamOptions) {
-    const timeout =
-      Platform.OS == 'android'
-        ? setTimeout(() => {
-            this.logger.log({
-              level: LogLevels.warn,
-              message: `HTTP Streaming POST is taking longer than ${Math.ceil(
-                STREAMING_POST_TIMEOUT_MS / 1000
-              )} seconds to resolve. If using a debug build, please ensure Flipper Network plugin is disabled.`
-            });
-          }, STREAMING_POST_TIMEOUT_MS)
-        : null;
-
-    try {
-      return await super.fetchStreamRaw({
-        ...options,
-        fetchOptions: {
-          ...options.fetchOptions,
-          /**
-           * The `react-native-fetch-api` polyfill provides streaming support via
-           * this non-standard flag
-           * https://github.com/react-native-community/fetch#enable-text-streaming
-           */
-          // @ts-expect-error https://github.com/react-native-community/fetch#enable-text-streaming
-          reactNative: { textStreaming: true }
-        }
-      });
-    } finally {
-      if (timeout) {
-        clearTimeout(timeout);
-      }
-    }
   }
 }
