+++ packages/shared-internals/src/client/sync/stream/AbstractRemote.ts
@@ -1,10 +1,8 @@
 import { type fetch } from 'cross-fetch';
-import { Requestable, RSocket, RSocketConnector } from 'rsocket-core';
 import PACKAGE from '../../../../package.json' with { type: 'json' };
 import { FetchStrategy, PowerSyncCredentials, LogLevels, PowerSyncLogger } from '@powersync/common';
 
 import { AbortOperation } from '../../../utils/AbortOperation.js';
-import { WebsocketClientTransport } from './WebsocketClientTransport.js';
 import {
   doneResult,
   extractBsonObjects,
@@ -12,7 +10,7 @@ import {
   SimpleAsyncIterator,
   valueResult
 } from '../../../utils/stream_transform.js';
-import { EventQueue } from '../../../utils/async.js';
+import { webSocketSyncStream } from './WebSocketSupport.js';
 
 /**
  * @internal
@@ -25,29 +23,13 @@ export type RemoteConnector = {
 const POWERSYNC_TRAILING_SLASH_MATCH = /\/+$/;
 const POWERSYNC_JS_VERSION = PACKAGE.version;
 
-const SYNC_QUEUE_REQUEST_HIGH_WATER = 10;
-const SYNC_QUEUE_REQUEST_LOW_WATER = 5;
-
-// Keep alive message is sent every period
-const KEEP_ALIVE_MS = 20_000;
-
-// One message of any type must be received in this period.
-const SOCKET_TIMEOUT_MS = 30_000;
-
-// One keepalive message must be received in this period.
-// If there is a backlog of messages (for example on slow connections), keepalive messages could be delayed
-// significantly. Therefore this is longer than the socket timeout.
-const KEEP_ALIVE_LIFETIME_MS = 90_000;
-
 /**
  * @internal
  */
 export type SyncStreamOptions = {
   path: string;
   data: unknown;
-  headers?: Record<string, string>;
   abortSignal: AbortSignal;
-  fetchOptions?: Request;
 };
 
 /**
@@ -55,46 +37,20 @@ export type SyncStreamOptions = {
  */
 export type FetchImplementation = typeof fetch;
 
-/**
- * Class wrapper for providing a fetch implementation.
- * The class wrapper is used to distinguish the fetchImplementation
- * option in [AbstractRemoteOptions] from the general fetch method
- * which is typeof "function"
- *
- * @internal
- */
-export class FetchImplementationProvider {
-  getFetch(): FetchImplementation {
-    throw new Error('Unspecified fetch implementation');
-  }
+export function lazyFetchImplementation(getFetch: () => FetchImplementation): FetchImplementation {
+  let resolved: FetchImplementation;
+
+  return (request) => {
+    const fetch = (resolved ??= getFetch());
+    return fetch(request);
+  };
 }
 
-/**
- * @internal
- */
-export type AbstractRemoteOptions = {
-  /**
-   * Transforms the PowerSync base URL which might contain
-   * `http(s)://` to the corresponding WebSocket variant
-   * e.g. `ws(s)://`
-   */
-  socketUrlTransformer: (url: string) => string;
-
-  /**
-   * Optionally provide the fetch implementation to use.
-   * Note that this usually needs to be bound to the global scope.
-   * Binding should be done before passing here.
-   */
-  fetchImplementation: FetchImplementation | FetchImplementationProvider;
-
-  /**
-   * Optional options to pass directly to all `fetch` calls.
-   *
-   * This can include fields such as `dispatcher` (e.g. for proxy support),
-   * `cache`, or any other fetch-compatible options.
-   */
-  fetchOptions?: {};
-};
+export interface FetchOptions {
+  resource: string;
+  request: RequestInit;
+  expectStreamingResponse: boolean;
+}
 
 /**
  * @internal
@@ -106,46 +62,27 @@ export interface SocketSyncStreamOptions {
   data: unknown;
 }
 
-/**
- * @internal
- */
-export const DEFAULT_REMOTE_OPTIONS: AbstractRemoteOptions = {
-  socketUrlTransformer: (url) =>
-    url.replace(/^https?:\/\//, function (match) {
-      return match === 'https://' ? 'wss://' : 'ws://';
-    }),
-  fetchImplementation: new FetchImplementationProvider(),
-  fetchOptions: {}
-};
+export interface PreparedRequest {
+  url: string;
+  headers: Record<string, string>;
+  userAgent: string;
+}
 
 /**
  * @internal
  */
 export abstract class AbstractRemote {
   protected credentials: PowerSyncCredentials | null = null;
-  protected options: AbstractRemoteOptions;
 
   constructor(
     protected connector: RemoteConnector,
-    protected logger: PowerSyncLogger,
-    options?: Partial<AbstractRemoteOptions>
-  ) {
-    this.options = {
-      ...DEFAULT_REMOTE_OPTIONS,
-      ...(options ?? {})
-    };
-  }
+    readonly logger: PowerSyncLogger
+  ) {}
 
   /**
-   * @returns a fetch implementation (function)
-   * which can be called to perform fetch requests
+   * Sends a request using a suitable `fetch` implementation for the current platform.
    */
-  get fetch(): FetchImplementation {
-    const { fetchImplementation } = this.options;
-    return fetchImplementation instanceof FetchImplementationProvider
-      ? fetchImplementation.getFetch()
-      : fetchImplementation;
-  }
+  protected abstract fetch(options: FetchOptions): Promise<Response>;
 
   /**
    * Get credentials currently cached, or fetch new credentials if none are
@@ -206,7 +143,7 @@ export abstract class AbstractRemote {
     return `powersync-js/${POWERSYNC_JS_VERSION}`;
   }
 
-  protected async buildRequest(path: string) {
+  protected async buildRequest(path: string): Promise<PreparedRequest> {
     const credentials = await this.getCredentials();
     if (credentials != null && (credentials.endpoint == null || credentials.endpoint == '')) {
       throw new Error('PowerSync endpoint not configured');
@@ -224,40 +161,23 @@ export abstract class AbstractRemote {
         'content-type': 'application/json',
         Authorization: `Token ${credentials.token}`,
         'x-user-agent': userAgent
-      }
-    };
-  }
-
-  async post(path: string, data: any, headers: Record<string, string> = {}): Promise<any> {
-    const request = await this.buildRequest(path);
-    const res = await this.fetch(request.url, {
-      method: 'POST',
-      headers: {
-        ...headers,
-        ...request.headers
       },
-      body: JSON.stringify(data)
-    });
-
-    if (res.status === 401) {
-      this.invalidateCredentials();
-    }
-
-    if (!res.ok) {
-      throw new Error(`Received ${res.status} - ${res.statusText} when posting to ${path}: ${await res.text()}}`);
-    }
-
-    return res.json();
+      userAgent
+    };
   }
 
   async get(path: string, headers?: Record<string, string>): Promise<any> {
     const request = await this.buildRequest(path);
-    const res = await this.fetch(request.url, {
-      method: 'GET',
-      headers: {
-        ...headers,
-        ...request.headers
-      }
+    const res = await this.fetch({
+      resource: request.url,
+      request: {
+        method: 'GET',
+        headers: {
+          ...headers,
+          ...request.headers
+        }
+      },
+      expectStreamingResponse: false
     });
 
     if (res.status === 401) {
@@ -279,7 +199,7 @@ export abstract class AbstractRemote {
     return new TextDecoder();
   }
 
-  protected createSocket(url: string): WebSocket {
+  createSocket(url: string): WebSocket {
     return new WebSocket(url);
   }
 
@@ -289,211 +209,7 @@ export abstract class AbstractRemote {
    * The only mechanism to abort the returned stream is to use the abort signal in {@link SocketSyncStreamOptions}.
    */
   async socketStreamRaw(options: SocketSyncStreamOptions): Promise<SimpleAsyncIterator<Uint8Array>> {
-    const { path, fetchStrategy = FetchStrategy.Buffered } = options;
-    const mimeType = 'application/json';
-
-    function toBuffer(js: any): Buffer {
-      return Buffer.from(JSON.stringify(js));
-    }
-
-    const syncQueueRequestSize = fetchStrategy == FetchStrategy.Buffered ? 10 : 1;
-    const request = await this.buildRequest(path);
-    const url = this.options.socketUrlTransformer(request.url);
-
-    // Add the user agent in the setup payload - we can't set custom
-    // headers with websockets on web. The browser userAgent is however added
-    // automatically as a header.
-    const userAgent = this.getUserAgent();
-
-    // While we're connecting (a process that can't be aborted in RSocket), the WebSocket instance to close if we wanted
-    // to abort the connection.
-    let pendingSocket: WebSocket | null = null;
-    let keepAliveTimeout: any;
-    let rsocket: RSocket | null = null;
-    let paused = false;
-    const queue = new EventQueue<Uint8Array | null>({
-      eventDelivered: () => {
-        if (queue.countOutstandingEvents <= SYNC_QUEUE_REQUEST_LOW_WATER) {
-          paused = false;
-          requestMore();
-        }
-      }
-    });
-    let didClose = false;
-    let connectionEstablished = false;
-    let pendingEventsCount = syncQueueRequestSize;
-    let res: Requestable | null = null;
-
-    const abortRequest = () => {
-      if (didClose) {
-        return;
-      }
-      didClose = true;
-
-      clearTimeout(keepAliveTimeout);
-
-      if (pendingSocket) {
-        pendingSocket.close();
-      }
-
-      if (rsocket) {
-        rsocket.close();
-      }
-
-      // Send a bogus event to the queue to ensure a pending listener gets woken up. We check for didClose and would
-      // return a doneEvent.
-      queue.notify(null);
-    };
-
-    function push(event: Uint8Array) {
-      queue.notify(event);
-      if (queue.countOutstandingEvents >= SYNC_QUEUE_REQUEST_HIGH_WATER) {
-        paused = true;
-      }
-    }
-
-    function requestMore() {
-      const delta = syncQueueRequestSize - pendingEventsCount;
-      if (!paused && delta > 0) {
-        res?.request(delta);
-        pendingEventsCount = syncQueueRequestSize;
-      }
-    }
-
-    // Handle upstream abort
-    if (options.abortSignal.aborted) {
-      throw new AbortOperation('Connection request aborted');
-    } else {
-      options.abortSignal.addEventListener('abort', abortRequest);
-    }
-
-    const resetTimeout = () => {
-      clearTimeout(keepAliveTimeout);
-      keepAliveTimeout = setTimeout(() => {
-        this.logger.log({
-          level: LogLevels.error,
-          message: `No data received on WebSocket in ${SOCKET_TIMEOUT_MS}ms, closing connection.`
-        });
-        abortRequest();
-      }, SOCKET_TIMEOUT_MS);
-    };
-    resetTimeout();
-
-    const connector = new RSocketConnector({
-      transport: new WebsocketClientTransport({
-        url,
-        wsCreator: (url) => {
-          const socket = (pendingSocket = this.createSocket(url));
-
-          socket.addEventListener('message', () => {
-            resetTimeout();
-          });
-          return socket;
-        }
-      }),
-      setup: {
-        keepAlive: KEEP_ALIVE_MS,
-        lifetime: KEEP_ALIVE_LIFETIME_MS,
-        dataMimeType: mimeType,
-        metadataMimeType: mimeType,
-        payload: {
-          data: null,
-          metadata: toBuffer({
-            token: request.headers.Authorization,
-            user_agent: userAgent
-          })
-        }
-      }
-    });
-
-    try {
-      rsocket = await connector.connect();
-      // The connection is established, we no longer need to monitor the initial timeout
-      pendingSocket = null;
-    } catch (ex) {
-      this.logger.log({ level: LogLevels.error, message: `Failed to connect WebSocket`, error: ex });
-      abortRequest();
-
-      throw ex;
-    }
-
-    resetTimeout();
-
-    // Helps to prevent double close scenarios
-    rsocket.onClose(() => (rsocket = null));
-
-    return await new Promise((resolve, reject) => {
-      const queueAsIterator: SimpleAsyncIterator<Uint8Array> = {
-        next: async () => {
-          if (didClose) return doneResult;
-
-          const notification = await queue.waitForEvent(options.abortSignal);
-          if (didClose) {
-            return doneResult;
-          } else {
-            return valueResult(notification!);
-          }
-        }
-      };
-
-      res = rsocket!.requestStream(
-        {
-          data: toBuffer(options.data),
-          metadata: toBuffer({
-            path
-          })
-        },
-        syncQueueRequestSize, // The initial N amount
-        {
-          onError: (e) => {
-            if (e.message.includes('PSYNC_')) {
-              if (e.message.includes('PSYNC_S21')) {
-                this.invalidateCredentials();
-              }
-            } else {
-              // Possible that connection is with an older service, always invalidate to be safe
-              if (e.message !== 'Closed. ') {
-                this.invalidateCredentials();
-              }
-            }
-
-            // Don't log closed as an error
-            if (e.message !== 'Closed. ') {
-              this.logger.log({ level: LogLevels.error, message: 'RSocket error', error: e });
-            }
-            // RSocket will close the RSocket stream automatically
-            // Close the downstream stream as well - this will close the RSocket connection and WebSocket
-            abortRequest();
-            // Handles cases where the connection failed e.g. auth error or connection error
-            if (!connectionEstablished) {
-              reject(e);
-            }
-          },
-          onNext: (payload) => {
-            // The connection is active
-            if (!connectionEstablished) {
-              connectionEstablished = true;
-              resolve(queueAsIterator);
-            }
-            const { data } = payload;
-
-            if (data) {
-              push(data);
-            }
-
-            // Less events are now pending
-            pendingEventsCount--;
-
-            // Request another event (unless the downstream consumer is paused).
-            requestMore();
-          },
-          onComplete: () => {
-            abortRequest(); // this will also emit a done event
-          },
-          onExtension: () => {}
-        }
-      );
-    });
+    return await webSocketSyncStream(this, options, await this.buildRequest(options.path));
   }
 
   /**
@@ -515,7 +231,7 @@ export abstract class AbstractRemote {
   protected async fetchStreamRaw(
     options: SyncStreamOptions
   ): Promise<{ isBson: boolean; stream: SimpleAsyncIterator<Uint8Array> }> {
-    const { data, path, headers, abortSignal } = options;
+    const { data, path, abortSignal } = options;
     const request = await this.buildRequest(path);
 
     /**
@@ -555,18 +271,19 @@ export abstract class AbstractRemote {
       const ndJson = 'application/x-ndjson';
       const bson = 'application/vnd.powersync.bson-stream';
 
-      res = await this.fetch(request.url, {
-        method: 'POST',
-        headers: {
-          ...headers,
-          ...request.headers,
-          accept: this.supportsStreamingBinaryResponses ? `${bson};q=0.9,${ndJson};q=0.8` : ndJson
+      res = await this.fetch({
+        resource: request.url,
+        request: {
+          method: 'POST',
+          headers: {
+            ...request.headers,
+            accept: this.supportsStreamingBinaryResponses ? `${bson};q=0.9,${ndJson};q=0.8` : ndJson
+          },
+          body: JSON.stringify(data),
+          signal: controller.signal,
+          cache: 'no-store'
         },
-        body: JSON.stringify(data),
-        signal: controller.signal,
-        cache: 'no-store',
-        ...(this.options.fetchOptions ?? {}),
-        ...options.fetchOptions
+        expectStreamingResponse: true
       });
 
       if (!res.ok || !res.body) {
