+++ packages/web/src/db/sync/WebRemote.ts
@@ -1,33 +1,18 @@
 import { LogLevels, PowerSyncLogger } from '@powersync/common';
-import {
-  AbstractRemote,
-  AbstractRemoteOptions,
-  FetchImplementation,
-  FetchImplementationProvider,
-  RemoteConnector
-} from '@powersync/shared-internals';
+import { AbstractRemote, FetchOptions, RemoteConnector } from '@powersync/shared-internals';
 
 import { getUserAgentInfo } from './userAgent.js';
-
-/*
- * Depends on browser's implementation of global fetch.
- */
-class WebFetchProvider extends FetchImplementationProvider {
-  getFetch(): FetchImplementation {
-    return fetch.bind(globalThis);
-  }
-}
 
 export class WebRemote extends AbstractRemote {
   constructor(
     protected connector: RemoteConnector,
-    protected logger: PowerSyncLogger,
-    options?: Partial<AbstractRemoteOptions>
+    logger: PowerSyncLogger
   ) {
-    super(connector, logger, {
-      ...(options ?? {}),
-      fetchImplementation: options?.fetchImplementation ?? new WebFetchProvider()
-    });
+    super(connector, logger);
+  }
+
+  protected fetch({ resource, request }: FetchOptions): Promise<Response> {
+    return fetch(resource, request);
   }
 
   getUserAgent(): string {
